# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kenya Administrative Boundaries API(KenyaAdminAPI) - Documentation",
        "description": "The Kenya Administrative Boundaries API (KenAdminAPI) provides a comprehensive and easy-to-use interface for accessing detailed information about Kenya's administrative divisions, including Counties, Constituencies, and Wards. This API is designed to help developers integrate location-based data into their applications, supporting a wide range of use cases from geographical analysis to service delivery.",
        "contact": {
            "email": "sosmongare@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        },
        {
            "url": "https://127.0.0.1:8000",
            "description": "Local secure server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/wards": {
            "get": {
                "tags": [
                    "Wards"
                ],
                "summary": "Get list of wards",
                "operationId": "fd963a232855fa6760db529d5a6d29c9",
                "responses": {
                    "200": {
                        "description": "Wards retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ward"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Wards"
                ],
                "summary": "Create a new ward",
                "operationId": "c1b6d51cc648a74b8b1772e91d16f0b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "constituency_id",
                                    "ward_name"
                                ],
                                "properties": {
                                    "constituency_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ward_name": {
                                        "type": "string",
                                        "example": "Parklands"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ward created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ward"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/wards/{id}": {
            "get": {
                "tags": [
                    "Wards"
                ],
                "summary": "Get a ward by ID",
                "operationId": "451b442228cb500e1c3ff98bfaeeffe6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ward retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ward"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Wards"
                ],
                "summary": "Update a ward",
                "operationId": "e49f9c42aae1c996a3399c47447c2317",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "constituency_id",
                                    "ward_name"
                                ],
                                "properties": {
                                    "constituency_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ward_name": {
                                        "type": "string",
                                        "example": "Parklands"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ward updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ward"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Wards"
                ],
                "summary": "Delete a ward",
                "operationId": "b84ab47cf19a93815411f8749ecc42fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ward deleted successfully"
                    }
                }
            }
        },
        "/api/wards/search/{query}": {
            "get": {
                "tags": [
                    "Wards"
                ],
                "summary": "Search for wards",
                "operationId": "b1ea501ba9098966042b9f8ae27b02ff",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ward"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No results found for the given query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Ward": {
                "title": "Ward",
                "description": "Ward model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the ward",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "constituency_id": {
                        "title": "Constituency ID",
                        "description": "ID of the constituency the ward belongs to",
                        "type": "integer",
                        "example": 1
                    },
                    "ward_name": {
                        "title": "Ward Name",
                        "description": "Name of the ward",
                        "type": "string",
                        "example": "Parklands"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Ward"
                }
            }
        }
    },
...
}